Refactor collection index moduke
