+++ ui/package.json
@@ -3,7 +3,7 @@
   "version": "3.5.1",
   "private": true,
   "dependencies": {
-    "@alephdata/followthemoney": "^1.25.5",
+    "@alephdata/followthemoney": "^1.26.0",
     "@alephdata/vislib": "^1.5.0",
     "@blueprintjs/core": "^3.18.1",
     "@blueprintjs/datetime": "^3.15.1",
